Move file zipping to utils
