+++ tools/figma-inspector/src/main.tsx
@@ -7,7 +7,6 @@ import {
     useCallback,
     useRef
 } from "react";
-import JSZip from "jszip";
 import {
     dispatchTS,
     listenTS,
@@ -17,21 +16,8 @@ import {
 import CodeSnippet from "./snippet/CodeSnippet";
 import "./main.css";
 import { useInspectorStore } from "./utils/store";
+import { downloadZipFile } from "./utils/utils.js";
 
-
-// Add file download functionality
-const downloadFile = (filename: string, text: string) => {
-    const element = document.createElement("a");
-    element.setAttribute(
-        "href",
-        "data:text/plain;charset=utf-8," + encodeURIComponent(text),
-    );
-    element.setAttribute("download", filename);
-    element.style.display = "none";
-    document.body.appendChild(element);
-    element.click();
-    document.body.removeChild(element);
-};
 
 export const App = () => {
     const [exportsAreCurrent, setExportsAreCurrent] = useState(false);
@@ -202,51 +188,7 @@ export const App = () => {
         return () => window.removeEventListener("message", directHandler);
     }, []);
 
-    // Create the functions with access to dispatchTS
-    const downloadZipFile = async (
-        files: Array<{ name: string; content: string }>,
-    ) => {
-        try {
-            if (!files || files.length === 0) {
-                console.error("No files to zip!");
-                return;
-            }
 
-            // Create a new JSZip instance directly (using the import)
-            const zip = new JSZip();
-
-            // Add each file to the zip with debug logging
-            files.forEach((file) => {
-                zip.file(file.name, file.content);
-            });
-
-            // Generate the zip
-            const content = await zip.generateAsync({ type: "blob" });
-
-            // Create download link
-            const element = document.createElement("a");
-            element.href = URL.createObjectURL(content);
-            element.download = "figma-collections.zip";
-            document.body.appendChild(element);
-            element.click();
-            document.body.removeChild(element);
-
-            // Clean up
-            URL.revokeObjectURL(element.href);
-        } catch (error) {
-            console.error("Error creating ZIP file:", error);
-
-            // Fallback to individual downloads if ZIP creation fails
-            alert(
-                "Couldn't create ZIP file. Downloading files individually...",
-            );
-            files.forEach((file, index) => {
-                setTimeout(() => {
-                    downloadFile(file.name, file.content);
-                }, index * 100);
-            });
-        }
-    };
 
     const buttonStyle: React.CSSProperties = {
         border: `none`,
