+++ tools/figma-inspector/src/utils/utils.ts
@@ -1,6 +1,6 @@
 // Copyright © SixtyFPS GmbH <info@slint.dev>
 // SPDX-License-Identifier: MIT
-import { dispatchTS } from "./bolt-utils.js";
+import JSZip from "jszip";
 
 export function writeTextToClipboard(str: string) {
     const prevActive = document.activeElement;
@@ -32,3 +32,59 @@ export function writeTextToClipboard(str: string) {
         }
     }
 }
+
+export const downloadZipFile = async (
+    files: Array<{ name: string; content: string }>,
+) => {
+    try {
+        if (!files || files.length === 0) {
+            console.error("No files to zip!");
+            return;
+        }
+
+        // Create a new JSZip instance directly (using the import)
+        const zip = new JSZip();
+
+        // Add each file to the zip with debug logging
+        files.forEach((file) => {
+            zip.file(file.name, file.content);
+        });
+
+        // Generate the zip
+        const content = await zip.generateAsync({ type: "blob" });
+
+        // Create download link
+        const element = document.createElement("a");
+        element.href = URL.createObjectURL(content);
+        element.download = "figma-collections.zip";
+        document.body.appendChild(element);
+        element.click();
+        document.body.removeChild(element);
+
+        // Clean up
+        URL.revokeObjectURL(element.href);
+    } catch (error) {
+        console.error("Error creating ZIP file:", error);
+
+        // Fallback to individual downloads if ZIP creation fails
+        alert("Couldn't create ZIP file. Downloading files individually...");
+        files.forEach((file, index) => {
+            setTimeout(() => {
+                downloadFile(file.name, file.content);
+            }, index * 100);
+        });
+    }
+};
+
+const downloadFile = (filename: string, text: string) => {
+    const element = document.createElement("a");
+    element.setAttribute(
+        "href",
+        "data:text/plain;charset=utf-8," + encodeURIComponent(text),
+    );
+    element.setAttribute("download", filename);
+    element.style.display = "none";
+    document.body.appendChild(element);
+    element.click();
+    document.body.removeChild(element);
+};
